Date And Time
The Calendar class is an abstract class that provides methods for converting between a specific instant in time and a set of calendar fields such as year, month, day of month, hour, and so on, and for manipulating the calendar fields, such as getting the date of the next week.

 
You are given a date. You just need to write the method, getDay, which returns the day on that date. To simplify your task, we have provided a portion of the code in the editor.

For example, if you are given the date August 14th 2017, the method should return MONDAY as the day on that date.