Basic Array Methods
You are given two integer arrays A and B, you have to delete first element of A and add that at the end of B and then append both the arrays. Finally, print the elements of merged array. For example: for A = [1, 2, 3] and B = [4,5], new array A = [2,3] and B = [4,5,1] and merged ans array = [2,3,4,5,1]. Please note you have to print one element of ans array in one line, so total outpur lines will be equal to the size of ans array.