JS Objects
You are given an Object mentioning marks of a student in three subjects. Please note 1st subject is Maths, 2nd is Physics and 3rd is Chemistry. All subject’s marks are distinct. Print the name of the Subject(Either Maths or Physics or Chemistry) having the highest marks. (You are not allowed to change the template.)