Class Inheritance
Given a class A having properties x and y. Declare a class B which inherits class A and has properties z and printIB (prints “IB”(without quotes) to the console when called). Also, the constructor of class B should set the values of x, y and z to the values provided in the user input.