Polymorphism
Given class Shape and two derived classes Square and Rectangle , all having area function as the only method. Define the area methods so that when area of a Shape is called, it always returns 0. While area of subclasses Square and Rectangle return area according to the dimensions provided as argument.