Vector
First line of the input contains an integer N denoting the number of integers.
Second line contains N space-separated integers. 
Third line contains an integer X denoting the index(0-based) of the N integers which needs to be deleted.

Your task is to print the N-1 integers in sorted order (ascending). Each integer in a new line.